backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
django==5.0.8; python_version >= '3.10'
django-ninja==1.2.1; python_version >= '3.7'
gunicorn==22.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
packaging==24.1; python_version >= '3.8'
psycopg[binary]==3.2.1; python_version >= '3.8'
psycopg-binary==3.2.1
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
rich==13.7.1; python_full_version >= '3.7.0'
sqlparse==0.5.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn==0.30.1; python_version >= '3.8'
uvicorn-worker==0.2.0; python_version >= '3.8'
